requirements.txt
Flask==3.0.2
SQLAlchemy==2.0.28
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
Flask-Login==0.6.3
WTForms==3.1.2
Flask-WTF==1.2.1
gunicorn==21.2.0
gevent==24.2.1
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5